genres uniquely associated with a show now is alerting the user after delete attempt

+++ Application/Services/GenreService.cs
@@ -7,13 +7,16 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.Services
 {
+   
     public class GenreService
     {
         private readonly GenreRepository _genreRepository;
+        private readonly ShowRepository _showRepository;
 
         public GenreService(ApplicationContext dbContext)
         {
             _genreRepository = new(dbContext);
+            _showRepository = new(dbContext);
         }
 
         public async Task<List<GenreViewModel>> GetAllViewModel()
@@ -55,10 +58,26 @@ namespace Application.Services
             await _genreRepository.AddAsync(genre);
         }
 
-        public async Task Delete(int id)
+
+        // Este method devuelve un booleano para saber qué mostrar en la vista dependiendo si el género
+        // a eliminar es o no único en una
+        public async Task<bool> Delete(int id)
         {
             var genre = await _genreRepository.GetByIdAsync(id);
+
+            // Aquí obtengo todas las series vinculadas con el Id del género que se quiere eliminar
+            var showsWithThisGenre = await _showRepository.GetShowsByGenreIdAsync(id);
+
+            // Verifico si alguna serie de la lista de arriba, tiene este género como UNICO género
+            bool hasShowsWithSingleGenre = showsWithThisGenre.Any(show => show.Genres.Count == 1);
+
+            if (hasShowsWithSingleGenre)
+            {
+                return false;
+            }
+
             await _genreRepository.DeleteAsync(genre);
+            return true;
         }
 
     }
